chore: Rename rag_config to llm_config
